ch4: move ch4r_comm functions into ch4r_proc
The functions in ch4r_comm.c has little to do with communicator but more
to do with av tables. Thus they are a bit misleading inside ch4r_comm.c.

Having them in ch4r_proc.c provides better context and easier to pruse
for critical details such access to global states and thread safety.
