+++ src/mpid/ch4/src/Makefile.mk
@@ -39,7 +39,6 @@ mpi_core_sources += src/mpid/ch4/src/ch4_globals.c        \
                     src/mpid/ch4/src/ch4i_comm.c          \
                     src/mpid/ch4/src/ch4r_buf.c           \
                     src/mpid/ch4/src/ch4r_init.c          \
-                    src/mpid/ch4/src/ch4r_comm.c          \
                     src/mpid/ch4/src/ch4r_proc.c          \
                     src/mpid/ch4/src/ch4r_recvq.c         \
                     src/mpid/ch4/src/ch4r_callbacks.c     \
