+++ src/mpid/ch4/src/ch4r_proc.h
@@ -22,6 +22,17 @@ int MPIDIU_get_max_node_id(MPIR_Comm * comm, int *max_id_p);
 int MPIDIU_build_nodemap(int myrank, MPIR_Comm * comm, int sz, int *out_nodemap, int *sz_out);
 int MPIDIU_build_nodemap_avtid(int myrank, MPIR_Comm * comm, int sz, int avtid);
 
+int MPIDIU_upids_to_lupids(int size, size_t * remote_upid_size, char *remote_upids,
+                           int **remote_lupids, int *remote_node_ids);
+int MPIDIU_Intercomm_map_bcast_intra(MPIR_Comm * local_comm, int local_leader, int *remote_size,
+                                     int *is_low_group, int pure_intracomm,
+                                     size_t * remote_upid_size, char *remote_upids,
+                                     int **remote_lupids, int *remote_node_ids);
+int MPIDIU_alloc_lut(MPIDI_rank_map_lut_t ** lut, int size);
+int MPIDIU_release_lut(MPIDI_rank_map_lut_t * lut);
+int MPIDIU_alloc_mlut(MPIDI_rank_map_mlut_t ** mlut, int size);
+int MPIDIU_release_mlut(MPIDI_rank_map_mlut_t * mlut);
+
 MPL_STATIC_INLINE_PREFIX int MPIDIU_comm_rank_to_pid(MPIR_Comm * comm, int rank, int *idx,
                                                      int *avtid)
 {
